refactor for consistent naming
