+++ web/src/beta/features/Editor/StoryPanel/PageIndicator/index.stories.tsx
