+++ web/src/beta/features/Editor/tabs/map/LeftPanel/GroupField/index.tsx
