+++ web/src/beta/features/Editor/tabs/story/LeftPanel/Action/index.stories.tsx
