+++ web/src/beta/features/Editor/tabs/story/LeftPanel/ContentPage/index.stories.tsx
