+++ web/src/beta/features/Editor/tabs/story/LeftPanel/ContentStory/index.stories.tsx
