+++ web/src/beta/features/Editor/tabs/story/LeftPanel/index.tsx
@@ -1,5 +1,5 @@
 import SidePanelCommon from "@reearth/beta/features/Editor/SidePanel";
-import ContentPage from "@reearth/beta/features/Editor/tabs/story/SidePanel/ContentPage";
+import ContentPage from "@reearth/beta/features/Editor/tabs/story/LeftPanel/ContentPage";
 import { StoryFragmentFragment, StoryPageFragmentFragment } from "@reearth/services/gql";
 import { useT } from "@reearth/services/i18n";
