+++ web/src/beta/features/Editor/tabs/widgets/RightPanel/Manager/Action.tsx
