+++ web/src/beta/features/Editor/tabs/widgets/RightPanel/Manager/index.tsx
