+++ web/src/beta/features/Editor/tabs/widgets/RightPanel/Settings/index.tsx
