+++ web/src/beta/features/Editor/tabs/widgets/RightPanel/index.tsx
